+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+### Added
+- Documentation for package plotter.
+
 ### Fixed
 - 'ng_help' no longer requires an argument for `-h`.